Change detection tweaks
