+++ replication/src/entities/components.rs
@@ -32,5 +32,8 @@ impl<T: Component> Plugin for ComponentReplicationPlugin<T> {
             fragmented: true,
             priority: self.message_priority,
         });
+
+        app.add_systems(PostUpdate, crate::change::undirty_components_system::<T>
+            .in_set(PostUpdateReplicationSystems::ClearDirty));
     }
 }
