+++ replication/src/scheduling.rs
@@ -24,4 +24,7 @@ pub enum PostUpdateReplicationSystems {
 
     /// Queues messages for sending.
     SendMessages,
+
+    /// Clears dirty flags.
+    ClearDirty,
 }
